Export node-postgres CommonJS module in addition to ES module (#92)
This PR adds a bundling step to the build, which produces a CommonJS
module in addition to the current ES module.

Without this, users that have a CommonJS package will get an error like
this when trying to import the package:

```
src/index.ts:1:34 - error TS1479: The current file is a CommonJS module whose imports will produce 'require' calls; however, the referenced file is an ECMAScript module and cannot be imported with 'require'. Consider writing a dynamic 'import("@aws/aurora-dsql-node-postgres-connector")' call instead.
  To convert this file to an ECMAScript module, change its file extension to '.mts' or create a local package.json file with `{ "type": "module" }`.

1 import { AuroraDSQLClient } from "@aws/aurora-dsql-node-postgres-connector";
                                   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Found 1 error in src/index.ts:1
```

While it is still possible to import the package with workarounds, it
adds a layer of friction for package adoption.

By adding `tsdown`, we can now run a bundling step which produces both
CommonJS and ES module files for the package.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.