Run node-postgres example smoke test as part of CI (#91)
This PR adds a CI step for `node-postgres` which checks the smoke test
in the example package.

While testing locally I found a few issues with running it, so this CI
job should help to prevent regressions.

The job uses `npm link` to reference the parent package directly, since
depending on the available version it may not yet be published on NPM,
and we want to test the code version described by the current commit.
For customers using this code sample the `package.json` still describes
a versioned dependency, so it will continue to work fine.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.