Configure Dependabot to monitor example projects (#97)
In #93 and #96, I added CI jobs to validate that 4 versions all matched:

- Package version
- Package lock version
- Example version
- Example lock version

While attempting to bump the version numbers in preparation for release,
I noticed we can update the first 3, but not the 4th until the package
is actually published to NPM, which breaks this workflow.

Instead, I've configured Dependabot to monitor the example packages,
which should automatically update the example versions once a new main
package version is released. This has a slight delay (~1 day unless
manually updated) but this way it won't be forgotten. It also has the
small benefit that anyone pulling the code sample from `main` in the
short window between when the versions are bumped and when the package
is published won't have to manually fix the version to run the example
code.

Given the changes in #91 and #95, the Dependabot change should be new
sufficient to make all 4 versions eventually consistent. The CI job
which verifies the version would need to be broken into 2 tasks which
verify the package versions and example versions separately, but this is
already handled by the new invocation of `npm ci` instead of `npm
install`. Given we invoke `npm ci` for both the package and the example,
any inconsistencies should surface as part of CI anyway, as can be seen
from invoking `npm ci` on 43e8bae03529d2c3f267f025f82e5d0c9e514980.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.