Remove broken Dependabot daily updates (#147)
We merged #146 in an attempt to configure Dependabot to upgrade
dependencies the way we wanted. As shown by [this failing
build](https://github.com/awslabs/aurora-dsql-nodejs-connector/runs/58587939940),
it is not possible to configure Dependabot with 2 different schedules
for the same package unfortunately.

We will need to find another way to keep the example versions
synchronized, but for now I'm removing the daily updates for the example
packages. This will keep the spam reduction from the original PR, but
will require us to manually synchronize the parent package versions, or
there will be a big lag between publishing.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.